Move several functions from utils to research/utils
